Additional.lyx: re-sync with branch
I don't know what happened with the version in master because it contained outdated sections I already removed months ago in branch. I can also not see new inf that is special in LyX 2.3dev.

Therefore I re-sync hereby with the version in branch.

As I often request: please change the doc files in branch, only new features of LyX 2.3 are master-only because there is no reason not to update and improve the docs for every LyX release.