Allow more translations of OutlinerNames
Allow comments at the end.

Although comments seem to be valid at the end of lines in layouts, few of the
regexps take this into account. However, OutlinerName was the only one where
there were actual lines with comments at the end. Not touching the others as
there is no one-size-fits-all solution with these regexps.

Thanks Kornel for reporting the issue.