aa.lyx: fix mixed layouts causing wrong PDF output
LyX correctly gave a warning about mixing InTitle layouts: There was
a LyX note in a Title environment, then there were standard
environments, and then a Title environment. This setup caused
several missing elements in the PDF.

Simply changing the Title environment of the Note to standard solved
the problems: The PDF output is now correct and LyX no longer gives
a warning.