Add some notes on forward/reverse search with evince.
Evince is a special case, since it provides f/r search not via command
line switches and pref settings, but via DBUS. On Linux at least, this
can be used by LyX via some external scripts. The possibility is now
mentioned here, the details (and the scripts) are provided on the wiki.